requirements.txt
anthropic>=0.92.0
python-frontmatter>=1.1.0
requests>=2.31.0
